Based the minion startup on pillar
This allows to run different version of this minion just for testing
purposes.

Added documentation to deploy another version on sorano.org. In this
state you still need to run the pre register with the correct names.

home_docker_minions.sls now has a list of dockers with versions and
names.
debian_minion.sls now uses the pillar to get the name and the version.